Install script is now more portable and part of the setup.

+++ install.sh
@@ -1,8 +1,18 @@
 # Zsh
-git clone git@github.com:agsimeonov/dotfiles.git ~/.dotfiles
-ln -s ~/.dotfiles/.zshenv ~/.
-ln -s ~/.dotfiles/.zshrc ~/.
-ln -s ~/.dotfiles/.zsh ~/.
+if ! [ -d ~/.dotfiles ]; then
+  git clone git@github.com:agsimeonov/dotfiles.git ~/.dotfiles
+fi
+if ! [ -e ~/.zshenv ]; then
+  ln -s ~/.dotfiles/.zshenv ~/.
+fi
+if ! [ -e ~/.zshrc ]; then
+  ln -s ~/.dotfiles/.zshrc ~/.
+fi
+if ! [ -e ~/.zsh ]; then
+  ln -s ~/.dotfiles/.zsh ~/.
+fi
 
 # Tmux
-ln -s ~/.dotfiles/.tmux.conf ~/.
+if ! [ -e ~/.tmux.conf ]; then
+  ln -s ~/.dotfiles/.tmux.conf ~/.
+fi
